First stab at the input side of external hint files. Only have CRC validation so far (almost).
Created a prefetching input stream to deal with cases where the OS prefetching is disabled to crank up the IOPs when an SSD is in play. Have not found out if it works.

Also snarf up a utility file for freeing direct byte buffers because the JVM sucks at it and they (for some value of they) have refused to acknowledge the problem for more than a decade. Yes I'm bitter.