Bend over bass ackwards to save a few bytes in the KeyDir.
Instead of using a separate byte array for keys and values use a single byte array. This saves a 4-byte length prefix (unless the JVM magically shrinks those) and in the future with a map implementation that doesn't suck it could use the same array reference for both key and value saving 4-8 bytes for the reference to the key/value.

Combine that with compressed object references and you are actually looking at a pretty space efficient implementation. Need to find a suitably licensed tree implementation to hack on.

Also stop storing the length of the value in the keydir saving another four bytes. Then length is stored with the value (and header) and is CRCed by both header CRC and value CRC so they can be evaluated separately without CRCing extra bytes. The tradeoff is a lot of complexity and an extra JNI call for reads to do the CRC. My hope is that the CRC implementation of the JVM will become an intrinsic.