Flip flopped AGAIN on how to store keys. This is simpler and does less work in the write thread. I did lose issuing timestamps from a single thread, need to fix that.
I decided I want to compress keys as well as values because it sucks paying the RAM cost for keys twice! Once in the page cache and one in the in memory index. If the index were directly memory mapped you wouldn't pay twice and this is with an approach that already sucks at storing large numbers of small keys.

I think I need to move the timestamp out of the compressed portion and issue them from the write thread to avoid having client threads see time move backwards.

I swear I will get to writing the hint file one of these days.